pyproject.toml
[tool.poetry]
name = "useful-habits"
version = "0.1.0"
description = ""
authors = ["mihasim <misha.dorozhkin@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.9"
djangorestframework-simplejwt = "^5.3.1"
requests = "^2.31.0"
celery = "^5.3.6"
redis = "^5.0.1"
django-cors-headers = "^4.3.1"
drf-yasg = "^1.21.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
